requirements.txt
python-keystoneclient
python-novaclient
flask
ujson
